Move like annotations to query helper
